pyproject.toml
[tool.poetry]
name = "Awesome-FastAPI-Boilerplate"
version = "0.1.0"
description = "An awesome FastaAPI boilerplate to rapidly get started with a well preprared backend microservice"
authors = ["Samuel Valdes Gutierrez"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
python-multipart = "^0.0.9"
alembic = "^1.13.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
uvicorn = {version = "^0.30.6", extras = ["standard"]}
python-jose = "^3.3.0"
sqlalchemy = "^2.0.32"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
# Pin bcrypt until passlib supports the latest
bcrypt = "4.0.1"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
mypy = "^1.11.1"
ruff = "^0.6.1"
pre-commit = "^3.8.0"
types-passlib = "^1.7.7.20240819"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
